add controllers, test, logging

+++ src/main/java/com/epam/cinema/service/AuditoriumServiceImpl.java
@@ -5,6 +5,9 @@ import com.epam.cinema.model.Auditorium;
 
 import java.util.List;
 
+import static java.util.Objects.isNull;
+import static org.springframework.util.Assert.notNull;
+
 public class AuditoriumServiceImpl implements AuditoriumService {
 
     AuditoriumDao auditoriumDao;
@@ -15,16 +18,25 @@ public class AuditoriumServiceImpl implements AuditoriumService {
 
     @Override
     public Long save(Auditorium object) {
+        notNull(object, "Value must not be null");
+
         return auditoriumDao.save(object);
     }
 
     @Override
     public void remove(Long id) {
-        auditoriumDao.remove(id);
+        notNull(id, "Value must not be null");
+
+        Auditorium auditorium = auditoriumDao.getById(id);
+        if (!isNull(auditorium)) {
+            auditoriumDao.remove(id);
+        }
     }
 
     @Override
     public Auditorium getById(Long id) {
+        notNull(id, "Value must not be null");
+
         return auditoriumDao.getById(id);
     }
 
@@ -35,6 +47,7 @@ public class AuditoriumServiceImpl implements AuditoriumService {
 
     @Override
     public Auditorium getByName(String name) {
+        notNull(name, "Value must not be null");
         return auditoriumDao.getByName(name);
     }
 }
